Move AllowUiaInChromium and AllowUiaInMSWord enums from UIAHandler to config and make use of AllowUiaInMSWord in profileUpgradeSteps.
