+++ source/UIAHandler/__init__.py
@@ -28,6 +28,10 @@ import threading
 import time
 import IAccessibleHandler.internalWinEventHandler
 import config
+from config import (
+	AllowUiaInChromium,
+	AllowUiaInMSWord,
+)
 import api
 import appModuleHandler
 import controlTypes
@@ -220,34 +224,6 @@ ignoreWinEventsMap = {
 }
 for id in UIAEventIdsToNVDAEventNames.keys():
 	ignoreWinEventsMap[id] = [0]
-
-
-class AllowUiaInChromium(Enum):
-	_DEFAULT = 0  # maps to 'when necessary'
-	WHEN_NECESSARY = 1  # the current default
-	YES = 2
-	NO = 3
-
-	@staticmethod
-	def getConfig() -> 'AllowUiaInChromium':
-		allow = AllowUiaInChromium(config.conf['UIA']['allowInChromium'])
-		if allow == AllowUiaInChromium._DEFAULT:
-			return AllowUiaInChromium.WHEN_NECESSARY
-		return allow
-
-
-class AllowUiaInMSWord(Enum):
-	_DEFAULT = 0  # maps to 'where suitable'
-	WHEN_NECESSARY = 1
-	WHERE_SUITABLE = 2
-	ALWAYS = 3
-
-	@staticmethod
-	def getConfig() -> 'AllowUiaInMSWord':
-		allow = AllowUiaInMSWord(config.conf['UIA']['allowInMSWord'])
-		if allow == AllowUiaInMSWord._DEFAULT:
-			return AllowUiaInMSWord.WHERE_SUITABLE
-		return allow
 
 
 def shouldUseUIAInMSWord(appModule: appModuleHandler.AppModule) -> bool:
